spark/requirements.txt
pyspark==3.5.3
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.3